Provision rearrangement
Installation functions get stored in their own file which the provision
script sources. The actual provision.sh script is gitignored because I
tend to destroy my VMs completely and rebuild them with only whatever
I'm interested in playing with at the time, and having to constantly
ignore lines commented in and out of the file when running git status is
annoying. Also added a gitignored sandbox directory to have it here
instead of living somewhere else.